pyproject.toml
[tool.poetry]
name = "ai-trading-agent"
version = "0.1.0"
description = "AI-Powered Trading Agent"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.13,<4.0"
ta-lib = ">=0.6.4,<0.7.0"
praw = ">=7.8.1,<8.0.0"
tweepy = ">=4.16.0,<5.0.0"
requests = ">=2.32.4,<3.0.0"
pandas = ">=2.3.1,<3.0.0"
yfinance = ">=0.2.65,<0.3.0"
vadersentiment = ">=3.3.2,<4.0.0"
pydantic = ">=2.11.7,<3.0.0"
fastapi = ">=0.116.1,<0.117.0"
uvicorn = ">=0.35.0,<1.0.0"
docker = ">=7.1.0,<8.0.0"
numpy = ">=2.3.0,<3.0"
python-dotenv = "^1.1.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
